requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
python-dotenv==1.0.0
schedule==1.2.0
colorama==0.4.6 